adding better documentation for bottomModal:
